--- components/calendar/src/datetime.rs
@@ -1,15 +1,0 @@
-// This file is part of ICU4X. For terms of use, please see the file
-// called LICENSE at the top level of the ICU4X source tree
-// (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
-
-use crate::{AsCalendar, Date, Time};
-
-/// A date + time for a given calendar.
-#[derive(Debug)]
-#[allow(clippy::exhaustive_structs)] // this type is stable
-pub struct DateTime<A: AsCalendar> {
-    /// The date
-    pub date: Date<A>,
-    /// The time
-    pub time: Time,
-}
