+++ components/datetime/examples/work_log.rs
@@ -9,9 +9,10 @@
 icu_benchmark_macros::instrument!();
 use icu_benchmark_macros::println;
 
-use icu_calendar::{Date, DateTime, Time};
+use icu_calendar::Date;
 use icu_datetime::{fieldsets::YMDT, FixedCalendarDateTimeFormatter};
 use icu_locale_core::locale;
+use icu_timezone::{DateTime, Time};
 
 const DATES_ISO: &[(i32, u8, u8, u8, u8, u8)] = &[
     (2001, 9, 8, 18, 46, 40),
