+++ components/datetime/src/pattern/pattern.rs
@@ -77,7 +77,7 @@ size_test!(DateTimePattern, date_time_pattern_size, 32);
 /// Check the hour cycle of a resolved pattern:
 ///
 /// ```
-/// use icu::calendar::Time;
+/// use icu::timezone::Time;
 /// use icu::datetime::provider::fields::components;
 /// use icu::datetime::fieldsets::T;
 /// use icu::datetime::pattern::DateTimePattern;
