+++ components/timezone/Cargo.toml
@@ -40,7 +40,7 @@ icu = { path = "../../components/icu", default-features = false }
 
 [features]
 default = ["compiled_data", "ixdtf"]
-ixdtf = ["dep:ixdtf"]
+ixdtf = ["dep:ixdtf", "icu_calendar/ixdtf"]
 serde = ["dep:serde", "zerovec/serde", "zerotrie/serde", "tinystr/serde", "icu_provider/serde"]
 datagen = ["serde", "dep:databake", "zerovec/databake", "zerotrie/databake", "tinystr/databake"]
 compiled_data = ["dep:icu_timezone_data", "icu_calendar/compiled_data"]
