+++ components/timezone/src/lib.rs
@@ -58,7 +58,8 @@
 //! # Examples
 //!
 //! ```
-//! use icu::calendar::{Date, Time};
+//! use icu::calendar::Date;
+//! use icu::timezone::Time;
 //! use icu::timezone::TimeZoneBcp47Id;
 //! use icu::timezone::TimeZoneIdMapper;
 //! use icu::timezone::ZoneVariant;
@@ -106,7 +107,8 @@ mod ids;
 pub mod provider;
 pub mod scaffold;
 mod time_zone;
-mod types;
+/// TODO
+pub mod types;
 mod windows_tz;
 mod zone_offset;
 
@@ -124,9 +126,9 @@ pub use provider::TimeZoneBcp47Id;
 pub use time_zone::models;
 pub use time_zone::TimeZoneInfo;
 pub use time_zone::TimeZoneModel;
-pub use types::{UtcOffset, ZoneVariant, ZonedDateTime};
+pub use types::{DateTime, Time, UtcOffset, ZoneVariant, ZonedDateTime};
 pub use windows_tz::{WindowsTimeZoneMapper, WindowsTimeZoneMapperBorrowed};
 pub use zone_offset::{ZoneOffsetCalculator, ZoneOffsets};
 
-#[cfg(all(feature = "ixdtf", feature = "compiled_data"))]
+#[cfg(feature = "ixdtf")]
 pub use crate::ixdtf::ParseError;
