+++ ffi/capi/bindings/js/Time.d.ts
@@ -6,7 +6,7 @@ import type { pointer, codepoint } from "./diplomat-runtime.d.ts";
 
 /** An ICU4X Time object representing a time in terms of hour, minute, second, nanosecond
 *
-*See the [Rust documentation for `Time`](https://docs.rs/icu/latest/icu/calendar/struct.Time.html) for more information.
+*See the [Rust documentation for `Time`](https://docs.rs/icu/latest/icu/timezone/struct.Time.html) for more information.
 */
 export class Time {
