+++ ffi/capi/src/datetime_formatter.rs
@@ -161,7 +161,7 @@ pub mod ffi {
             value: &IsoDateTime,
             write: &mut diplomat_runtime::DiplomatWrite,
         ) {
-            let greg = icu_calendar::DateTime {
+            let greg = icu_timezone::DateTime {
                 date: icu_calendar::Date::new_from_iso(value.0.date, icu_calendar::Gregorian),
                 time: value.0.time,
             };
@@ -220,7 +220,7 @@ pub mod ffi {
             value: &IsoDateTime,
             write: &mut diplomat_runtime::DiplomatWrite,
         ) {
-            let greg = icu_calendar::DateTime {
+            let greg = icu_timezone::DateTime {
                 date: icu_calendar::Date::new_from_iso(value.0.date, icu_calendar::Gregorian),
                 time: value.0.time,
             };
@@ -312,7 +312,7 @@ pub mod ffi {
             value: &IsoDateTime,
             write: &mut diplomat_runtime::DiplomatWrite,
         ) -> Result<(), DateTimeFormatError> {
-            let any = icu_calendar::DateTime {
+            let any = icu_timezone::DateTime {
                 date: value.0.date.to_any(),
                 time: value.0.time,
             };
@@ -387,7 +387,7 @@ pub mod ffi {
             value: &IsoDateTime,
             write: &mut diplomat_runtime::DiplomatWrite,
         ) -> Result<(), DateTimeFormatError> {
-            let any = icu_calendar::DateTime {
+            let any = icu_timezone::DateTime {
                 date: value.0.date.to_any(),
                 time: value.0.time,
             };
