+++ ffi/capi/src/time.rs
@@ -12,13 +12,13 @@ pub mod ffi {
 
     #[diplomat::opaque]
     /// An ICU4X Time object representing a time in terms of hour, minute, second, nanosecond
-    #[diplomat::rust_link(icu::calendar::Time, Struct)]
-    pub struct Time(pub icu_calendar::Time);
+    #[diplomat::rust_link(icu::timezone::Time, Struct)]
+    pub struct Time(pub icu_timezone::Time);
 
     impl Time {
         /// Creates a new [`Time`] given field values
-        #[diplomat::rust_link(icu::calendar::Time::try_new, FnInStruct)]
-        #[diplomat::rust_link(icu::calendar::Time::new, FnInStruct, hidden)]
+        #[diplomat::rust_link(icu::timezone::Time::try_new, FnInStruct)]
+        #[diplomat::rust_link(icu::timezone::Time::new, FnInStruct, hidden)]
         #[diplomat::attr(supports = fallible_constructors, constructor)]
         pub fn create(
             hour: u8,
@@ -30,7 +30,7 @@ pub mod ffi {
             let minute = minute.try_into()?;
             let second = second.try_into()?;
             let nanosecond = nanosecond.try_into()?;
-            let time = icu_calendar::Time {
+            let time = icu_timezone::Time {
                 hour,
                 minute,
                 second,
@@ -40,42 +40,42 @@ pub mod ffi {
         }
 
         /// Creates a new [`Time`] from an IXDTF string.
-        #[diplomat::rust_link(icu::calendar::Time::try_from_str, FnInStruct)]
-        #[diplomat::rust_link(icu::calendar::Time::try_from_utf8, FnInStruct, hidden)]
-        #[diplomat::rust_link(icu::calendar::Time::from_str, FnInStruct, hidden)]
+        #[diplomat::rust_link(icu::timezone::Time::try_from_str, FnInStruct)]
+        #[diplomat::rust_link(icu::timezone::Time::try_from_utf8, FnInStruct, hidden)]
+        #[diplomat::rust_link(icu::timezone::Time::from_str, FnInStruct, hidden)]
         #[diplomat::attr(supports = fallible_constructors, named_constructor)]
         pub fn from_string(v: &DiplomatStr) -> Result<Box<Time>, CalendarParseError> {
-            Ok(Box::new(Time(icu_calendar::Time::try_from_utf8(v)?)))
+            Ok(Box::new(Time(icu_timezone::Time::try_from_utf8(v)?)))
         }
 
         /// Creates a new [`Time`] representing midnight (00:00.000).
-        #[diplomat::rust_link(icu::calendar::Time::midnight, FnInStruct)]
+        #[diplomat::rust_link(icu::timezone::Time::midnight, FnInStruct)]
         #[diplomat::attr(supports = fallible_constructors, named_constructor)]
         pub fn midnight() -> Result<Box<Time>, CalendarError> {
-            let time = icu_calendar::Time::midnight();
+            let time = icu_timezone::Time::midnight();
             Ok(Box::new(Time(time)))
         }
 
         /// Returns the hour in this time
-        #[diplomat::rust_link(icu::calendar::Time::hour, StructField)]
+        #[diplomat::rust_link(icu::timezone::Time::hour, StructField)]
         #[diplomat::attr(auto, getter)]
         pub fn hour(&self) -> u8 {
             self.0.hour.into()
         }
         /// Returns the minute in this time
-        #[diplomat::rust_link(icu::calendar::Time::minute, StructField)]
+        #[diplomat::rust_link(icu::timezone::Time::minute, StructField)]
         #[diplomat::attr(auto, getter)]
         pub fn minute(&self) -> u8 {
             self.0.minute.into()
         }
         /// Returns the second in this time
-        #[diplomat::rust_link(icu::calendar::Time::second, StructField)]
+        #[diplomat::rust_link(icu::timezone::Time::second, StructField)]
         #[diplomat::attr(auto, getter)]
         pub fn second(&self) -> u8 {
             self.0.second.into()
         }
         /// Returns the nanosecond in this time
-        #[diplomat::rust_link(icu::calendar::Time::nanosecond, StructField)]
+        #[diplomat::rust_link(icu::timezone::Time::nanosecond, StructField)]
         #[diplomat::attr(auto, getter)]
         pub fn nanosecond(&self) -> u32 {
             self.0.nanosecond.into()
