+++ provider/source/src/time_zones/convert.rs
@@ -10,10 +10,10 @@ use cldr_serde::time_zones::time_zone_names::*;
 use core::cmp::Ordering;
 use icu::calendar::Date;
 use icu::calendar::Iso;
-use icu::calendar::Time;
 use icu::datetime::provider::time_zones::*;
 use icu::locale::{langid, LanguageIdentifier};
 use icu::timezone::provider::*;
+use icu::timezone::Time;
 use icu::timezone::UtcOffset;
 use icu::timezone::ZoneVariant;
 use icu_provider::prelude::*;
